+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * New interactive command `clojure-view-reference-section`.
 * New interactive command `clojure-view-cheatsheet`.
 * New interactive command `clojure-view-grimoire`.
+* New interactive command `clojure-view-style-guide`.
 * Make the refactoring keymap prefix customizable via `clojure-refactor-map-prefix`.
 
 ## 5.5.2 (2016-08-03)